% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildH.R
\name{buildH}
\alias{buildH}
\title{Spatial dependence matrix of the factor loadings}
\usage{
buildH(areapoly, permutation = NA)
}
\arguments{
\item{areapoly}{The polygon of the areas. We can obtain this through \code{readOGR} function from \code{sp} matrix.}

\item{permutation}{Permutation order of the subregions}
}
\value{
A list of two matrices: Precision matrix H and the covariance matrix obtained through Moore-Penrose inverse of H.
}
\description{
It computes the spatial covariance and precision matrix of the neighboring subregions using Intrinsice Autoregressive Conditional (ICAR) process.
}
\details{
The off-digonal values are -1 when two subregions are neighbors. Otherwise, we assign 0. The diagonal values are the sum of the values of its own row.
}
