\name{DIFM-package}
\alias{DIFM-package}
\alias{DIFM}
\docType{package}
\title{
\packageTitle{DIFM}
}
\description{
\packageDescription{DIFM}
}
\details{
Package: BCFM2

Type: Package

Version: 1.0

Date: 2023-02-20

License: GPL(>=2)
}
\author{
\packageAuthor{DIFM}

Maintainer: \packageMaintainer{DIFM}
}
\references{
Shin, H. and Ferreira, M. (2023). "Dynamic ICAR Spatiotemporal Factor Models." Spatial Statistics, 56, 100763

Lopes, H. and West, M. (2004). “Bayesian Model Assessment in Factor Analysis.” Statistica Sinica, 14, 41–67.

Prado, R., Ferreira, M. A. R., and West, M. (2021). Time Series: Modeling, Computation, and Inference. 2nd ed. Boca Raton: CRC Press.
}
\keyword{ package }