% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/difm.model.attributes.R
\name{difm.model.attributes}
\alias{difm.model.attributes}
\title{Initialize model attributes for DIFM}
\usage{
difm.model.attributes(data, n.iter, n.factors, G0)
}
\arguments{
\item{data}{The dataset}

\item{n.iter}{Number of iterations}

\item{n.factors}{Number of factors to run DIFM}

\item{G0}{The basic evolution matrix for one factor}
}
\value{
A list of number of timepoints, subregions, factors, matrix of evolution matrix, and matrix to extract common factors.
}
\description{
It initialize the basic parameters and model attributes for DIFM
}
