% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DIFMcpp.R
\name{DIFMcpp}
\alias{DIFMcpp}
\title{Run Dynamic ICAR Factors Model (DIFM), with C++ codes}
\usage{
DIFMcpp(model.attributes, hyp.parm, data, every = 1, verbose = TRUE)
}
\arguments{
\item{model.attributes}{Model attributes from \code{difm.model.attributes}}

\item{hyp.parm}{Hyperparameters from \code{difm.hyp.parm}}

\item{data}{The dataset}

\item{every}{Save \code{every} iterations to final result}

\item{verbose}{Print out the iteration process}
}
\value{
The Gibbs sampler of DIFM
}
\description{
This function runs Dynamic ICAR factors Model (DIFM), simulated from C++ codes
}
