% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal.d.R
\name{marginal.d}
\alias{marginal.d}
\title{Marginal predictive density}
\usage{
marginal.d(
  data,
  model.attributes,
  hyp.parm,
  Gibbs,
  burnin = NA,
  verbose = TRUE
)
}
\arguments{
\item{data}{The dataset}

\item{model.attributes}{Model attributes generated from \code{difm.model.attributes}.}

\item{hyp.parm}{Hyperparameters generated from \code{difm.hyp.parm}.}

\item{Gibbs}{Result of Gibbs sampler from DIFM function.}

\item{burnin}{Burn-in period. If not specified, one tenths of the iterations will be the burn-in period.}

\item{verbose}{Print out the iteration process.}
}
\value{
Metropolis-Laplace estimator of the Marginal density
}
\description{
It calculates the marginal density (Lewis and Raftery, 1997) from the DIFM sample using R.
}
