% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/permutation.order.R
\name{permutation.order}
\alias{permutation.order}
\title{Order of permutation by the largest absolute value in each eigenvector}
\usage{
permutation.order(data, n.factors)
}
\arguments{
\item{data}{The dataset}

\item{n.factors}{Number of factors}
}
\value{
The numeric vector of permutation
}
\description{
It finds the vector of permutation to permute data by its largest absolute value in each eigenvector. It sets the order by specified number of factors, and the rest is ordered as they were.
}
