% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_B.CI.R
\name{plot_B.CI}
\alias{plot_B.CI}
\title{Credible interval plot of factor loadings}
\usage{
plot_B.CI(
  Gibbs,
  true.val = NA,
  burnin = NA,
  permutation = NA,
  main.bool = TRUE,
  layout.dim = NA
)
}
\arguments{
\item{Gibbs}{Result of Gibbs sampler from DIFM function}

\item{true.val}{True values of factor loadings. If not available, NA.}

\item{burnin}{Number of burn-in. If not specified, it uses the first tenths as burn-in period.}

\item{permutation}{Permutation of variables. If not specified, no permutation.}

\item{main.bool}{Add title of the plots.}

\item{layout.dim}{Dimension of panel layout for multiple factor loadings. If not specificed, factor loadings plots are layout in one column.}
}
\value{
Factor loadings credible interval plots
}
\description{
The functions builds a column-wise plots of factor loadings. The parameters fixed at 1 are displayed with red dashed vertical lines.
}
