% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sigma2.CI.R
\name{plot_sigma2.CI}
\alias{plot_sigma2.CI}
\title{A credible interval plot of posterior of sigma squared}
\usage{
plot_sigma2.CI(Gibbs, burnin = NA, permutation = NA, main.bool = TRUE)
}
\arguments{
\item{Gibbs}{Result of Gibbs sampler from DIFM function.}

\item{burnin}{Number of burn-in. If not specified, it uses the first tenths as burn-in period.}

\item{permutation}{Permutation of variables. If not specified, no permutation.}

\item{main.bool}{Add title of the plots.}
}
\value{
A credible interval plot of sigma squared
}
\description{
It returns a credible interval plot of idiosyncratic variance, sigma squared. The lines are 95% intervals, while the circles are posterior mean.
}
