% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_tau.CI.R
\name{plot_tau.CI}
\alias{plot_tau.CI}
\title{Credible interval plot of factor loadings variance}
\usage{
plot_tau.CI(Gibbs, burnin = NA, true.val = NA, main.bool = TRUE)
}
\arguments{
\item{Gibbs}{Result of Gibbs sampler from DIFM function.}

\item{burnin}{Number of burn-in. If not specified, it uses the first tenths as burn-in period.}

\item{true.val}{True values of tau. If not available, NA.}

\item{main.bool}{Add title of the plots.}
}
\value{
Credible interval plot of tau
}
\description{
It returns a credible interval plot of factor loadings covariance, tau. The lines are 95% intervals, while the circles are posterior mean.
}
