% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_X.CI.R
\name{plot_X.CI}
\alias{plot_X.CI}
\title{Credible interval plot of common factors}
\usage{
plot_X.CI(Gibbs, burnin = NA, main.bool = FALSE, layout.dim = NA)
}
\arguments{
\item{Gibbs}{Result of Gibbs sampler from DIFM function.}

\item{burnin}{Number of burn-in. If not specified, it uses the first tenths as burn-in period.}

\item{main.bool}{Add title of the plots.}

\item{layout.dim}{Dimension of panel layout for multiple common factors. If not specificed, common factor plots are layout in one column.}
}
\value{
Credible interval plots of common factors
}
\description{
The functions builds the plot of 95\% confidence intervals of the common realizations, X. The black solid lines are the posterior mean and the dased lines are the 95\% confidence intervals.
}
