% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Property}
\alias{Property}
\title{Property crime in United States}
\format{
## `Property`
A data frame with 60 rows and 11 columns:
\describe{
  \item{AZ}{Arizona}
  \item{CA}{California}
  \item{CO}{Colorado}
  \item{ID}{Idaho}
  \item{MT}{Montana}
  \item{NV}{Nevada}
  \item{NM}{New Mexico}
  \item{OR}{Oregon}
  \item{UT}{Utah}
  \item{WA}{Washington}
  \item{WY}{Wyoming}
  ...
}
}
\source{
<https://www.disastercenter.com/crime/>
}
\usage{
Property
}
\description{
A subset of data of property crime per 100,000 people in western states from 1960 to 2019.
}
\keyword{datasets}
