% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{WestStates}
\alias{WestStates}
\title{Westen states in United States}
\format{
## `WestStates`
A SpatialPolygonsDataFrame data of the western states in United States
\describe{
  \item{FID}{The number ID of the western states}
  \item{State_Code}{Abbreviations of the state names}
  \item{State_Name}{Names of the states}
  A SpatialPolygonsDataFrame data of the western states in United States
}
}
\source{
<https://www.census.gov/geographies/mapping-files/time-series/geo/carto-boundary-file.html>
}
\usage{
WestStates
}
\description{
A sp map data of the western states in United States
}
\keyword{datasets}
